Added Crud repository working model to add database and perform basic operations.

+++ mmproject2-scratchcopy/src/main/java/com/fil/mmproject/repositories/AdminRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.fil.mmproject.models.Admin;
 
 @Repository
-public interface AdminRepo extends JpaRepository<Admin, Integer>{
+public interface AdminRepo extends JpaRepository<Admin, Long>{
 
 }
